backend/pyproject.toml
[project]
name = "swarmsight-backend"
version = "0.1.0"
description = "SwarmSight AI - Multi-Agent Social Simulation Engine"
requires-python = ">=3.11,<3.13"
dependencies = [
    "flask>=3.0",
    "flask-cors>=4.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "camel-oasis>=0.2.5",
]

[tool.uv]
dev-dependencies = []
